context is first chars, use uint32 instead of ulong

+++ src/markov/markov_model.h
@@ -23,7 +23,7 @@ private:
     double smoothingFactor;
     unsigned int reduceFactor;
     unsigned int alphabetSize;
-    unordered_map<string, unordered_map<char, unsigned long>> table;
+    unordered_map<string, unordered_map<char, unsigned int>> table;
 
     void saveTableToCache();
